Use size_t for maxjobs and maxfail

+++ build.h
@@ -1,7 +1,7 @@
 struct node;
 
 struct buildoptions {
-	int maxjobs, maxfail;
+	size_t maxjobs, maxfail;
 	_Bool verbose, explain;
 };
